Stats Learning Course
This is a statistical learning course with a focus on supervised and
unsupervised learning. It includes regression and classification
methods.Also, includes: linear and polynomial regression, logistic
regression and linear discriminant analysis; cross-validation and the
bootstrap, model selection and regularization methods (ridge and lasso);
nonlinear models, splines and generalized additive models; tree-based
methods, random forests and boosting; support-vector machines. Some
unsupervised learning methods are discussed: principal components and
clustering (k-means and hierarchical).

The lectures cover all the material in An Introduction to Statistical
Learning, with Applications in R by James, Witten, Hastie and Tibshirani
(Springer, 2013). As of January 5, 2014, the pdf for this book will be
available for free, with the consent of the publisher, on the book
website.

URL:
https://class.stanford.edu/courses/HumanitiesandScience/StatLearning/Winter2015/about